Moved FunctionConfig/SourceConfig/SinkConfig to a more common location
