+++ pulsar-common/src/main/java/org/apache/pulsar/common/functions/WindowConfig.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.functions.utils;
+package org.apache.pulsar.common.functions;
 
 import lombok.Data;
 import lombok.Getter;
