+++ pulsar-common/src/main/java/org/apache/pulsar/common/io/SinkConfig.java
@@ -16,9 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.functions.utils;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+package org.apache.pulsar.common.io;
 
 import java.util.Collection;
 import java.util.Map;
@@ -29,63 +27,40 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
+import org.apache.pulsar.common.functions.ConsumerConfig;
+import org.apache.pulsar.common.functions.FunctionConfig;
+import org.apache.pulsar.common.functions.Resources;
 
 @Getter
 @Setter
 @Data
 @EqualsAndHashCode
 @ToString
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class FunctionConfig {
-
-    public enum ProcessingGuarantees {
-        ATLEAST_ONCE,
-        ATMOST_ONCE,
-        EFFECTIVELY_ONCE
-    }
-
-    public enum Runtime {
-        JAVA,
-        PYTHON
-    }
+public class SinkConfig {
 
     private String tenant;
     private String namespace;
     private String name;
     private String className;
-    private Collection<String> inputs;
-    private Map<String, String> customSerdeInputs;
-    private String topicsPattern;
-    private Map<String, String> customSchemaInputs;
+    private String sourceSubscriptionName;
 
-    /**
-     * A generalized way of specifying inputs
-     */
+    private Collection<String> inputs;
+
+    private Map<String, String> topicToSerdeClassName;
+
+    private String topicsPattern;
+
+    private Map<String, String> topicToSchemaType;
+
     private Map<String, ConsumerConfig> inputSpecs = new TreeMap<>();
 
-    private String output;
-
-    /**
-     * Represents either a builtin schema type (eg: 'avro', 'json', ect) or the class name for a Schema
-     * implementation
-     */
-    private String outputSchemaType;
-
-    private String outputSerdeClassName;
-    private String logTopic;
-    private ProcessingGuarantees processingGuarantees;
-    private boolean retainOrdering;
-    private Map<String, Object> userConfig;
-    private Runtime runtime;
-    private boolean autoAck;
-    private int maxMessageRetries = -1;
-    private String deadLetterTopic;
-    private String subName;
+    private Map<String, Object> configs;
     private int parallelism = 1;
+    private FunctionConfig.ProcessingGuarantees processingGuarantees;
+    private boolean retainOrdering;
     private Resources resources;
-    private String fqfn;
-    private WindowConfig windowConfig;
+    private boolean autoAck;
     private Long timeoutMs;
-    private String jar;
-    private String py;
+
+    private String archive;
 }
