+++ pulsar-common/src/main/java/org/apache/pulsar/common/io/SourceConfig.java
@@ -16,48 +16,39 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.functions.utils;
-
-import java.util.Collection;
-import java.util.Map;
-import java.util.TreeMap;
+package org.apache.pulsar.common.io;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
+import org.apache.pulsar.common.functions.FunctionConfig;
+import org.apache.pulsar.common.functions.Resources;
+
+import java.util.Map;
 
 @Getter
 @Setter
 @Data
 @EqualsAndHashCode
 @ToString
-public class SinkConfig {
-
+public class SourceConfig {
     private String tenant;
     private String namespace;
     private String name;
     private String className;
-    private String sourceSubscriptionName;
 
-    private Collection<String> inputs;
+    private String topicName;
 
-    private Map<String, String> topicToSerdeClassName;
+    private String serdeClassName;
 
-    private String topicsPattern;
-
-    private Map<String, String> topicToSchemaType;
-
-    private Map<String, ConsumerConfig> inputSpecs = new TreeMap<>();
+    private String schemaType;
 
     private Map<String, Object> configs;
     private int parallelism = 1;
     private FunctionConfig.ProcessingGuarantees processingGuarantees;
-    private boolean retainOrdering;
     private Resources resources;
-    private boolean autoAck;
-    private Long timeoutMs;
 
     private String archive;
 }
